rename: kafka 관련 설정을 /commom/kafka 패키지로 이동
