+++ src/main/java/com/skyhorsemanpower/payment/common/kafka/KafkaProducerCluster.java
@@ -1,14 +1,10 @@
-package com.skyhorsemanpower.payment.common;
+package com.skyhorsemanpower.payment.common.kafka;
 
 import java.util.concurrent.CompletableFuture;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.kafka.core.KafkaTemplate;
-import org.springframework.kafka.support.KafkaHeaders;
 import org.springframework.kafka.support.SendResult;
-import org.springframework.messaging.Message;
-import org.springframework.messaging.support.MessageBuilder;
 import org.springframework.stereotype.Component;
 
 @Slf4j
